services/image/imggen/requirements.txt
flask==3.0.0
torch==2.0.1
torchvision==0.15.2
diffusers==0.21.4
transformers==4.30.2
accelerate==0.20.3
pillow==10.0.1
numpy==1.24.3
huggingface_hub==0.16.4
